updated tables migration file

+++ back-end/src/db/migrations/20241111130814_tables.js
@@ -14,5 +14,9 @@ exports.up = function (knex) {
   };
   
   exports.down = function (knex) {
-    return knex.schema.dropTable("tables");
+    return knex.schema.hasTable("tables").then((exists) => {
+      if (exists) {
+        return knex.schema.dropTable("tables");
+      }
+    });
   };
